requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
openai>=1.0.0
psutil>=5.9.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
line_profiler>=4.0.0
memory-profiler>=0.61.0
regex>=2023.0.0
pydantic>=2.0.0
pytest>=7.0.0
jupyter>=1.0.0
python-dotenv>=1.0.0